refactor(home): querify home exclu offer
