+++ src/features/home/components/modules/exclusivity/helpers/useShouldDisplayExcluOffer.ts
@@ -1,4 +1,4 @@
-import { useExcluOffer } from 'features/home/api/useExcluOffer'
+import { useExcluOfferQuery } from 'features/home/queries/useExcluOfferQuery'
 import { ExclusivityModule } from 'features/home/types'
 import { getOfferPrice } from 'features/offer/helpers/getOfferPrice/getOfferPrice'
 import { useMaxPrice } from 'features/search/helpers/useMaxPrice/useMaxPrice'
@@ -12,7 +12,7 @@ export function useShouldDisplayExcluOffer(
   const { userLocation } = useLocation()
 
   const maxPrice = useMaxPrice()
-  const { data: offer } = useExcluOffer(offerId)
+  const { data: offer } = useExcluOfferQuery(offerId)
 
   if (!offer) return false
